Fixed n_elements being a double in all parameters file - should be unsigned. Added some comments to algebra file

+++ one_d_demag_testing/parameters-demag3.cc
@@ -12,7 +12,7 @@ namespace OneDMicromagSetup
   double dt = 0.05;
 
   // Number of elements
-  double n_x_elements = 40;
+  unsigned n_x_elements = 40;
 
   // Prototypes for coeff functions
   double llg_damping_coeff(const double& t, const Vector<double>& x);
