+++ one_d_demag_testing/parameters-demag6.cc
@@ -14,7 +14,7 @@ namespace OneDMicromagSetup
   double dt = 0.05;
 
   // Number of elements
-  double n_x_elements = 41;
+  unsigned n_x_elements = 41;
 
   // Prototypes for coeff functions
   double llg_damping_coeff(const double& t, const Vector<double>& x);
